scripts/requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
boto3>=1.28.0
